chore: move tests from WinnerTest to LottoNumbersTest
